+++ crypto/include/aes_icm.h
@@ -56,7 +56,7 @@ typedef struct {
     v128_t keystream_buffer;              /* buffers bytes of keystream       */
     srtp_aes_expanded_key_t expanded_key; /* the cipher key                   */
     int bytes_in_buffer;                  /* number of unused bytes in buffer */
-    int key_size;                         /* AES key size + 14 byte SALT */
+    size_t key_size;                      /* AES key size + 14 byte SALT */
 } srtp_aes_icm_ctx_t;
 
 #endif /* AES_ICM_H */
