+++ test/util.h
@@ -44,10 +44,15 @@
 #ifndef SRTP_TEST_UTIL_H
 #define SRTP_TEST_UTIL_H
 
+#include <stddef.h>
+
 #define MAX_PRINT_STRING_LEN 1024
 
-int hex_string_to_octet_string(char *raw, char *hex, int len);
-char *octet_string_hex_string(const void *s, int length);
-int base64_string_to_octet_string(char *raw, int *pad, char *base64, int len);
+size_t hex_string_to_octet_string(char *raw, char *hex, size_t len);
+char *octet_string_hex_string(const void *s, size_t length);
+size_t base64_string_to_octet_string(char *raw,
+                                     int *pad,
+                                     char *base64,
+                                     size_t len);
 
 #endif
